pyproject.toml
[tool.poetry]
name = "allin"
version = "0.1.0"
description = "Allin is an experimental asynchronous web framework."
authors = ["Aprila Hijriyan <hijriyan23@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
typing-extensions = "^4.2.0"
autoroutes = "^0.3.5"
biscuits = "^0.3.0"
msgspec = "^0.12.0"
fast-query-parsers = "^0.3.0"
jetblack-asgi-typing = "^0.4.0"
python-multipart = "^0.0.5"
aiofiles = "^22.1.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.18.1"
uvicorn = {version = "^0.20.0", extras = ["standard"]}
python-semantic-release = "^7.33.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_toml = "pyproject.toml:tool.poetry.version"

[tool.ruff]
exclude = [
    "examples/",
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    ".venv",
]
target-version = "py39"
select = ["F", "E", "W", "B", "Q", "UP", "C90", "N", "S", "A", "C4", "PIE", "T20", "SIM"]
fix = true
line-length = 120
show-source = true
unfixable = []
ignore = ["S101"]

[tool.ruff.pydocstyle]
convention = "google"
